missing pagination on /wallet page, also missing dispensers section & remove border of carousel image &  perhaps some duplication in islands/Carousel.tsx vs islands/Carousel3d.tsx that needs to be resolved
